+++ CHANGELOG.md
@@ -98,3 +98,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/jamcgrath/svelte-component-visualizer/releases/tag/0.2.0
 [0.2.1]: https://github.com/jamcgrath/svelte-component-visualizer/releases/tag/v0.2.1
 [0.3.0]: https://github.com/jamcgrath/svelte-component-visualizer/releases/tag/v0.3.0
+[0.4.0]: https://github.com/jamcgrath/svelte-component-visualizer/releases/tag/v0.4.0